[Agent, UI]: Added a definition for agent tables and abstract handlers
Defined a table factory to automatically generate tables based on handlers
and their states. Added a thread manager to only threads created by handlers.

This changes allow for the creation of new classes with ease, the following
changes must aim to use this architecture to add the next two agents.